requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.1
dj-config-url==0.1.1
Django==3.1
django-cors-headers==3.2.1
django-dramatiq==0.9.1
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dramatiq==1.9.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
prometheus-client==0.8.0
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.12.0
pytz==2019.3
regex==2020.7.14
requests==2.24.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.45.0
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
